# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  http.DelegationCreateRequest:
    properties:
      delegate_id:
        type: string
      ends_at:
        type: string
      reason:
        type: string
      scope:
        type: string
      scope_id:
        type: string
      starts_at:
        type: string
    type: object
  http.DelegationListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/http.DelegationResponse'
        type: array
    type: object
  http.DelegationResponse:
    properties:
      delegate_id:
        type: string
      delegator_id:
        type: string
      ends_at:
        type: string
      id:
        type: string
      reason:
        type: string
      record_version:
        type: integer
      scope:
        type: string
      scope_id:
        type: string
      starts_at:
        type: string
      status:
        type: string
    type: object
...
  http.ErrorResponse:
    properties:
      active_workflows:
        description: |-
          ActiveWorkflows is populated on 409 workflow_resolution_required
          responses (§8.8). Count of open workflows blocking the mutation.
        example: 3
        type: integer
      allowed_actions:
        description: AllowedActions lists the P-26 resolution actions the caller may
          invoke.
        example:
        - replace_delegate
        - stop_workflows
        items:
          type: string
        type: array
      code:
        description: Code is a legacy alias for Error kept for backwards compatibility.
        example: not_found
        type: string
      details:
        description: Details is populated with per-field errors for 422 validation
          failures.
        items:
          $ref: '#/definitions/http.ValidationError'
        type: array
      error:
        description: Error is the machine-readable error code (LLD §17 error taxonomy).
        example: not_found
        type: string
      message:
        description: Message is a human-readable explanation of the error.
        example: resource not found
        type: string
      record_version:
        description: |-
          RecordVersion is populated on 409 optimistic_lock_conflict responses
          so the client can re-read and retry (CONC-4).
        example: 5
        type: integer
      request_id:
        description: RequestID is the per-request UUID injected by platform-gincommon.
        example: 01H8XYZ0000000000000000000
        type: string
      status:
        description: Status is the HTTP status code echoed in the body.
        example: 404
        type: integer
      trace_id:
        description: TraceID is the OTel W3C trace ID for cross-service correlation.
        example: a3f1b2c8d4e5f60718293a4b5c6d7e8f
        type: string
      workflow_ids:
        description: WorkflowIDs is populated on 409 workflow_resolution_required
          responses.
        items:
          type: string
        type: array
    type: object
...
  http.TenderACLGrantRequest:
    properties:
      access_level:
        type: string
      expires_at:
        type: string
      reason:
        type: string
      user_id:
        type: string
    type: object
  http.TenderACLListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/http.TenderACLResponse'
        type: array
    type: object
  http.TenderACLResponse:
    properties:
      access_level:
        type: string
      expires_at:
        type: string
      granted_by:
        type: string
      reason:
        type: string
      user_id:
        type: string
    type: object
  http.TenderACLRevokeResponse:
    properties:
      revoked:
        example: true
        type: boolean
      user_id:
        format: uuid
        type: string
    type: object
  http.ValidationError:
    properties:
      code:
        description: Code is the machine-readable violation code.
        example: out_of_range
        type: string
      field:
        description: Field is the JSON path of the failing field.
        example: mfa_freshness_seconds
        type: string
      message:
        description: Message is a human-readable explanation of the field violation.
        example: must be between 60 and 900
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: sharmila.dayalan@bcbpsolutions.com
    name: BCBP Solutions
  description: |-
    Organizational layer of the IAM subsystem — owns tenants, departments, memberships, tenant-level roles, delegations, invitations, ACL overlays, and group→role mappings.

    **Tenant isolation.** Every resource is scoped by `x-tenant-id`; cross-tenant reads are blocked by row-level security (RLS-1..RLS-6). The hot path `GET /internal/users/:id/memberships` is the enrichment source for AuthZ (SLO 15 ms cache-hit / 30 ms cache-miss).

    **Route prefixes.**  `/api/v1/*` — authenticated tenant users (JWT via gateway).  `/api/v1/internal/*` — in-mesh services (mTLS, `iam-system` role).  `/api/v1/operator/*` — platform operators (`platform_operator` role).

    Every mutation echoes `record_version` for optimistic-locking round-trip (CONC-4).
  license:
    name: Proprietary
  title: IAM Org & Membership API
  version: "1.0"
paths:
  /delegations:
    get:
      description: Returns active delegations for the caller's tenant (RLS-scoped).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.DelegationListResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: P-18 — List active delegations
      tags:
      - delegations
    post:
      consumes:
      - application/json
      description: §8.6 availability-first — calls UP SetAvailability BEFORE inserting
        delegations row. On UP failure returns 422 invalid_delegate.
      parameters:
      - description: Delegation payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.DelegationCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.DelegationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "422":
          description: invalid_delegate | self_delegation | delegation_window_inverted
            | scope_id_required
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: P-19 — Create OOO delegation (self-service, AUTH-4)
      tags:
      - delegations
  /delegations/{id}:
    delete:
      description: '§8.7 pointer-clear: calls UP with {delegate_id: null} first, then
        flips status to ''cancelled''. Self-service; admin can cancel any.'
      parameters:
      - description: Delegation UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Optimistic-lock version
        in: query
        name: record_version
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.DelegationResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: record_version mismatch (CONC-4)
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: P-20 — Cancel delegation
      tags:
      - delegations
...
  /tenants/{id}/tenders/{tender_id}/acl:
    get:
      description: AUTH-3 tender_admin/tenant_admin/owner. Returns view/edit/approve
        grants that have not expired.
      parameters:
      - description: Tenant UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Tender UUID
        format: uuid
        in: path
        name: tender_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.TenderACLListResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: P-21 — List active ACL entries for a tender
      tags:
      - acl
    post:
      consumes:
      - application/json
      description: AUTH-3 tender_admin/tenant_admin/owner. Optional expires_at (future-only).
      parameters:
      - description: Tenant UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Tender UUID
        format: uuid
        in: path
        name: tender_id
        required: true
        type: string
      - description: Grant payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.TenderACLGrantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.TenderACLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "422":
          description: invalid_expires_at | invalid_access_level
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: P-22 — Grant tender ACL (view/edit/approve)
      tags:
      - acl
  /tenants/{id}/tenders/{tender_id}/acl/{user_id}:
    delete:
      description: AUTH-3 tender_admin/tenant_admin/owner. Soft-deletes the ACL row.
      parameters:
      - description: Tenant UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Tender UUID
        format: uuid
        in: path
        name: tender_id
        required: true
        type: string
      - description: User UUID
        format: uuid
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.TenderACLRevokeResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: P-23 — Revoke tender ACL
      tags:
      - acl
securityDefinitions:
  TenantID:
    description: Tenant UUID injected by the API gateway. Establishes the RLS scope
      for the request (RLS-6).
    in: header
    name: x-tenant-id
    type: apiKey
  TenantRoles:
    description: Comma-separated tenant-role list injected by the API gateway. Operator
      routes require `platform_operator`; internal routes require `iam-system`.
    in: header
    name: x-tenant-roles
    type: apiKey
  UserID:
    description: Authenticated user UUID injected by the API gateway. Use `iam-system`
      for in-mesh service calls hitting `/api/v1/internal/*`.
    in: header
    name: x-user-id
    type: apiKey
swagger: "2.0"